Refactor lua code to return rule status
Instead of returning a number to indicate whether a given request has
violated a rate limit, return a JSON-encoded string with some state
including the current number of requests in a given bucket as well as
whether or not that bucket is violated. This starts the refactoring to
make remaining requests/reset information available via response
headers.